sandbox - more lifting of extensions related management contributions
